Eliminate directory creation confusion and patch CreateDirectoryA.
The existing calls to win32_utf8's custom CreateDirectory function assumed that the directory part ended at the last (back)slash. This, however, is not how CreateDirectory is supposed to operate - it interprets the *entirety* of the string as a directory name.

This fixes CreateDirectoryU to work as intended, and creates a new function, dir_create_for_fn, which should be used if the old behavior is desired.

With a fixed function, we can now do IAT patching for CreateDirectoryU. This is indeed necessary for games from th125 on -  the initial creation of the games' AppData directory otherwise would fail if a user name contains non-ASCII characters.