win32_utf8: Add DragQueryFile().
This came out of a fruitless attempt to turn EditDisk into Unicode for an
easier setup of the PC-98 English patches - which, of course, could have never
worked because we *must* write Shift-JIS names to the disk in the end, and we
wouldn't have been able to pull that off without huge binary hacks.

Whatever, this code's supposed to work.