Add WritePrivateProfileString().
Eh, the Touhou-Online launcher is too hard. It actually does everything right,
but it can't be recompiled without UNICODE defined, which would be required in
order for it to work with win32_utf8. It's also not exactly trivial to rewrite
it that way, since it interfaces with libxml2 (UTF-8 internally), GDI+
(UTF-16) and all the embedded resources. Since this is only supposed to be a
side project next to ReC98, let's cover FMPMDE instead. ^_^

... oh, hey, yet another function requiring a 50-line hack in order to
transparently output actual Unicode.