Delay-load all DLLs that might not need to be loaded into the target process.
Not only is this bound to reduce program startup time a bit, which will become
more and more important the larger win32_utf8 becomes, it also eliminates
"DLL not found" errors if win32_utf8 is loaded via LoadLibraryEx() and a flag
that excludes the Windows system directories from the search path.

Unsurprisingly, GCC and MinGW have no equivalent for this very useful feature.
:P