w32u8_strcmp: assert() the pointers instead of comparing them.
Not only because it was actually wrong to do so (it previously returned
"strings are equal" if one of the pointers was NULL, which is certainly
not what you want), but also because some people might actually need
the precise non-zero result.

Asserts seem to be a good compromise though, which neither the MSVCRT
nor MinGW do by default.