Define wmain() in addition to main().
Saving those precious nanoseconds that the C runtime would otherwise
waste with needlessly converting the command line from UTF-16 to the
ANSI codepage. Y'know, I figured it might raise questions if I omit
this, and maybe you don't want to have that feel of tainting your build
with an unnecessary ANSI function…

But yeah, entry point precedence *was* a question I had when writing
these wrappers, so at least it serves as documentation now. But yeah,
main() ultimately *is* the superior choice, due to both being
cross-platform and requiring less build configuration.

Now, if GCC had a #pragma to control its entry point selection behavior,
it would be awesome and I'd be gushing with praise at this marvelous
piece of compiler technology. This is still pretty nice though.