Properly supports strings with mixed encoding
Some strings can end up with a mix of local encoding and UTF-8 in them.
The most common case is when the program gets an UTF-8 string from a
win32_utf8 function, and then uses sprintf (or strcat or whatever) to combine
it with a string that did not go through win32_utf8 and it still in the local
encoding. With such string, an UTF-8 -> UTF-16 conversion will fail, and
win32_utf8 will fall back to convert it from the fallback encoding to UTF-16,
destroying the UTF-8 text in it in the process.

With this commit, when such text is detected during the conversion, we convert
each code point individually with their proper source encoding.